Fixed concurrency bug
